Define JobAmmendmentFactory as a test service (#419)
